Cache + pagination + authentication

+++ src/Controller/UserController.php
@@ -13,7 +13,8 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
-use Symfony\Component\Serializer\SerializerInterface ;
+use JMS\Serializer\SerializationContext;
+use JMS\Serializer\SerializerInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 class UserController extends AbstractController
@@ -21,7 +22,18 @@ class UserController extends AbstractController
     #[Route('api/users/{id}', name: 'user_list', methods: ['GET'])]
     public function getUserList(Customer $customer, SerializerInterface $serializer): JsonResponse
     {
-        $jsonUsers = $serializer->serialize($customer->getUsers(), 'json', ['groups' => 'getUsers']);
+        // Checking access (customer can only access his own user list)
+        $loggedCustomer = $this->getUser();
+        if ($loggedCustomer->getId() != $customer->getId()) {
+            $response = $serializer->serialize([
+                'status' => '401',
+                'message' => 'Invalid credentials.',
+            ], 'json');
+            return new JsonResponse($response, Response::HTTP_UNAUTHORIZED, ['accept' => 'json'], true);
+        }
+
+        $context = SerializationContext::create()->setGroups(['getUsers']);
+        $jsonUsers = $serializer->serialize($customer->getUsers(), 'json', $context);
         return new JsonResponse($jsonUsers, Response::HTTP_OK, ['accept' => 'json'], true);
     }
 
@@ -32,20 +44,39 @@ class UserController extends AbstractController
         UserRepository $userRepository, SerializerInterface $serializer
         ): JsonResponse
     {
-        // TODO Authentication
+        // Checking access (customer can only access his own user list)
         $customer = $customerRepository->find($customerId);
+        $loggedCustomer = $this->getUser();
+        if ($loggedCustomer->getId() != $customer->getId()) {
+            $response = $serializer->serialize([
+                'status' => '401',
+                'message' => 'Invalid credentials.',
+            ], 'json');
+            return new JsonResponse($response, Response::HTTP_UNAUTHORIZED, ['accept' => 'json'], true);
+        }
+
+        // TODO check if user is not empty
         $user = $userRepository->find($userId);
-        $jsonUsers = $serializer->serialize($user, 'json', ['groups' => 'getUsers']);
+        $context = SerializationContext::create()->setGroups(['getUsers']);
+        $jsonUsers = $serializer->serialize($user, 'json', $context);
         return new JsonResponse($jsonUsers, Response::HTTP_OK, ['accept' => 'json'], true);
     }
 
     #[Route('api/users/{customerId}/{userId}', name: 'delete_user', methods: ['DELETE'])]
     public function deleteUser(
         int $customerId, int $userId, EntityManagerInterface $em, CustomerRepository $customerRepository, 
-        UserRepository $userRepository): JsonResponse
+        UserRepository $userRepository, SerializerInterface $serializer): JsonResponse
     {
-        // TODO Authentication
+        // Checking access (customer can only access his own user list)
         $customer = $customerRepository->find($customerId);
+        $loggedCustomer = $this->getUser();
+        if ($loggedCustomer->getId() != $customer->getId()) {
+            $response = $serializer->serialize([
+                'status' => '401',
+                'message' => 'Invalid credentials.',
+            ], 'json');
+            return new JsonResponse($response, Response::HTTP_UNAUTHORIZED, ['accept' => 'json'], true);
+        }
         $user = $userRepository->find($userId);
         $em->remove($user);
         $em->flush();
@@ -59,6 +90,16 @@ class UserController extends AbstractController
         EntityManagerInterface $em, UrlGeneratorInterface $urlGenerator,
         ValidatorInterface $validator): JsonResponse
     {
+        // Checking access (customer can only access his own user list)
+        $loggedCustomer = $this->getUser();
+        if ($loggedCustomer->getId() != $customer->getId()) {
+            $response = $serializer->serialize([
+                'status' => '401',
+                'message' => 'Invalid credentials.',
+            ], 'json');
+            return new JsonResponse($response, Response::HTTP_UNAUTHORIZED, ['accept' => 'json'], true);
+        }
+
         $user = $serializer->deserialize($request->getContent(), User::class, 'json');
 
         $errors = $validator->validate($user);
@@ -75,7 +116,8 @@ class UserController extends AbstractController
         $em->persist($user);
         $em->flush();
 
-        $jsonUser = $serializer->serialize($user, 'json', ['groups' => 'getUsers']);
+        $context = SerializationContext::create()->setGroups(['getUsers']);
+        $jsonUser = $serializer->serialize($user, 'json', $context);
         $location = $urlGenerator->generate('user_details', [
             'customerId' => $customer->getId(),
             'userId' => $user->getId()
